Make `no_tx_action` configurable
There are cases where a user could want to configure the `no_tx_action`
behavior of `before_commit` and `after_commit`.

For example, I have a case where I want to capture a Stripe payment in
an `after_commit` callback. It's important that the API call happen
after other changes have been committed to the database to avoid
charging a customer and then rolling back and leaving our database in a
state where it looks like a customer hasn't paid their bill. In cases
like this I'd prefer to use the `:exception` setting so that running our
test suite on CI will raise if we have a call path that doesn't include
an open transaction.